Contact rate calculation for wild deer

Contact rate calculation for wild deer

Usage

calc_contact_rate(
  sigma_season = NULL,
  scaling_c = NULL,
  N_w = NULL,
  q = NULL,
  A_w = NULL,
  nsamples = NULL,
  seed = NULL,
  type_contact = NULL
)

Arguments

sigma_season

value for season scaling (derived from Williams et al 2014; Appendix A). Default set to 1

scaling_c

value for contact scaling constant (influences slope of density-contact relationship; reported in Habib et al. 2011; Appendix A)

N_w

total population size of deer. Defaults is a random draw from a Poisson dist w/mean = 1000

q

value for concavity scaling constant (0-1, with 0 equating to density dependence and 1 equating to frequency dependence; reported in Habib et al. 2011, Appendix A)

A_w

area inhabited by N. Default set to 100

nsamples

number of values to be returned. Default is 1, but will return a vector if > 2

seed

if specified, sets a seed for the function

type_contact

option between "low", "medium", and "high" based on parameters from Habib 2011 table. If set to "manual", manual input of scaling_c and q are needed.

Value

returns a number or numeric vector if nsamples > 1

Examples

calc_contact_rate(type_contact = "low")
#> [1] 4.297626
calc_contact_rate(sigma_season = 1, scaling_c = 16.37, N_w = 1000, q = 0.53, A_w = 100)
#> [1] "Type of contact must be specified as 'low', 'med', 'high'. If set to 'manual', user must provide scaling_c and q values "
calc_contact_rate(c(1,1), c(16.37, 16.37), rpois(2, 1000), c(0.53, 0.53), c(100, 100))
#> [1] "Type of contact must be specified as 'low', 'med', 'high'. If set to 'manual', user must provide scaling_c and q values "