Calculate probability of infection via aerosol \(\nu^{AERO}\)

calc_nu_aero will calculate one value for probability of infection using defaults if no arguments are provided. Defaults are described in the sir_model_description document and are sourced from the literature or expert elicitation.

Usage

calc_nu_aero(
  AER = NULL,
  s = NULL,
  lambda = NULL,
  C_nu = NULL,
  C_i = NULL,
  IR = NULL,
  ER = NULL,
  V_d = NULL,
  V_air = NULL,
  t_contact = NULL,
  r = NULL,
  nsamples = NULL,
  seed = NULL
)

Arguments

AER

air exchange in \(hr^{-1}\). Default is set to 4.

s

settling rate; \(hr^{-1}\) Default is set to 0.24

lambda

innactivation rates \(hr^{-1}\). Default is set to 0.63

C_nu

viral load in sputum; RNA copies/ml. Default samples from expert elicited distribution of parameter 'Viral Load'

C_i

conversion factor for quanta/RNA copy. Default set to 0.0014

IR

inhalation rate; \(m^3/hr\).Default is set to 0.846

ER

exhalation rate; \(m^3/hr\). Default IR = ER

V_d

exhaled droplet volume concentration; ml exhaled droplets/ \(m^3\). Default to 0.009

V_air

fixed volume; \(m^3\). Default to 7.07

t_contact

time of contact with contaminated airspace (hr).

r

species-specific probability of infection from 1 quantum. Default is for r_deer with expert elicited values.

nsamples

default to 1, but if specified > 1 will draw nsamples from the default distributions of parameters

seed

if setting a seed, specify number

Value

a numeric vector of length equal to nsamples for values of nu_aero

Details

Mathematical background for this calculation. An infected individual emits viral particles at a particular rate \(ER_q\) in quanta/hr as the product of the arguments described above: $$ER_q = C_{\nu} \cdot C_i \cdot ER \cdot V_d$$ This is used to model the instantaneous concentration of viral particles (C) in well-mixed air space (quanta/\(m^3\)) around an infected individual as follows: $$C = \frac{ER_q}{IVRR \cdot V_{air}}$$ where the loss rate (IVRR) is given by: $$AER + s + \lambda$$ When a susceptible individual enters the contaminated airspace surrounding an infected individual, the dose (\(Q_A\)) is the product of inhalation rate, concentration of viral particles, and time of contact: $$Q_A = IR \cdot C \cdot t_{contact}$$ The dose \(Q_A\) is converted into a probability of infection using the Wells-Riley infection model as a function of the dose received and a species-specific probability of infection from 1 quantum. $$\nu^{AERO} = 1 - e^{-rQ}$$

Examples

calc_nu_aero() # will return a single value using defaults
#> [1] 0.006387472