Contact rate parameters

Modeling the relationship between density and contact rates in wild deer populations. Habib et al. (2011) used radio-collared deer in eastern Alberta to, in part, estimate contact rates between deer in the same group and deer in different groups. These results are unique as they calculate the per capita contact rate in terms of deer contacted/individual/time. This differs from many other approaches, which just look at the raw probability of two collared individuals being in the same place at the same time. Habib et al. (2011) estimated home ranges, contact rates, and resource selection functions across variable densities and habitat conditions (areas with 12%, 26%, 53%, and 67% wooded habitat). They then simulated individual white-tailed deer and their movements based on these observed data to estimate the total number of individuals in contact, which was the sum of contact rates within the same group and with different groups. They estimated the most likely parameters for the per-capita contact rate.

Usage

contact_rate_params

Format

contact_rate_params

Source

https://doi.org/10.1016/j.ecolmodel.2011.05.007