Random draw of parameters from expert elicitation

Random draw of parameters from expert elicitation

Usage

draw_elicitation_samples(
  elicitation_data = NULL,
  nsamples = NULL,
  seed = NULL,
  return_df = FALSE
)

Arguments

elicitation_data

defaults to the package data on expert elicitation. It requires a dataframe with columns for parameter name, mu, sd, and distribution family. This function is written around the elicitation_data dataset provided in this package.

nsamples

a number of samples to draw. If none set, it defaults to 1.

seed

if specified, a seed is set. If left NULL no seed is set.

return_df

if TRUE the object returned is a data frame. If FALSE it will return a list column

Value

returns a tibble with list column for the samples drawn unless otherwise specified in the return_df argument.

Examples

draw_elicitation_samples() # this will return the default elicitation_data with an additional column for a random sample of 1 for each parameter.
#> # A tibble: 15 × 5
#>    parameter                                           mu    sd family my_sample
#>    <chr>                                            <dbl> <dbl> <chr>  <list>   
#>  1 Temporary Immunity                              4.72   0.626 log-n… <dbl [1]>
#>  2 Wastewater Infections                          -1.35   1.49  log-n… <dbl [1]>
#>  3 Fomite Infections                              -0.0414 1.03  log-n… <dbl [1]>
#>  4 Viral Load                                      0.216  0.344 log-n… <dbl [1]>
#>  5 Dose-Response                                   0.277  0.272 log-n… <dbl [1]>
#>  6 Deer Proximity Duration (minutes)               1.55   1.27  log-n… <dbl [1]>
#>  7 Direct Contact Probability                     -1.46   0.708 logit… <dbl [1]>
#>  8 Proximity rate with baiting (17 events withou…  3.47   0.226 log-n… <dbl [1]>
#>  9 Deer-Human Proximity Rate, Rural (per 120 day… -1.59   1.70  log-n… <dbl [1]>
#> 10 Deer-Human Proximity Duration, Rural (minutes) -0.355  0.979 log-n… <dbl [1]>
#> 11 Deer-Human Proximity Rate, Suburban (per 120 …  0.573  0.951 log-n… <dbl [1]>
#> 12 Deer-Human Proximity Duration, Suburban (minu…  0.432  0.929 log-n… <dbl [1]>
#> 13 Deer-Human Proximity Rate, Captive (per 120 d…  2.52   1.13  log-n… <dbl [1]>
#> 14 Deer-Human Proximity Duration, Captive (minut…  1.79   1.15  log-n… <dbl [1]>
#> 15 Deer-Deer Proximity Rate, Captive (per day)     3.47   0.913 log-n… <dbl [1]>
draw_elicitation_samples(return_df = TRUE) # will return the same as above but expanded as a dataframe instead of a tibble with a list column
#> # A tibble: 15 × 5
#>    parameter                                           mu    sd family my_sample
#>    <chr>                                            <dbl> <dbl> <chr>      <dbl>
#>  1 Temporary Immunity                              4.72   0.626 log-n…   74.3   
#>  2 Wastewater Infections                          -1.35   1.49  log-n…    0.0765
#>  3 Fomite Infections                              -0.0414 1.03  log-n…    0.917 
#>  4 Viral Load                                      0.216  0.344 log-n…    1.14  
#>  5 Dose-Response                                   0.277  0.272 log-n…    1.48  
#>  6 Deer Proximity Duration (minutes)               1.55   1.27  log-n…    2.00  
#>  7 Direct Contact Probability                     -1.46   0.708 logit…    0.140 
#>  8 Proximity rate with baiting (17 events withou…  3.47   0.226 log-n…   26.2   
#>  9 Deer-Human Proximity Rate, Rural (per 120 day… -1.59   1.70  log-n…    0.0925
#> 10 Deer-Human Proximity Duration, Rural (minutes) -0.355  0.979 log-n…    1.50  
#> 11 Deer-Human Proximity Rate, Suburban (per 120 …  0.573  0.951 log-n…    2.86  
#> 12 Deer-Human Proximity Duration, Suburban (minu…  0.432  0.929 log-n…    1.38  
#> 13 Deer-Human Proximity Rate, Captive (per 120 d…  2.52   1.13  log-n…   27.6   
#> 14 Deer-Human Proximity Duration, Captive (minut…  1.79   1.15  log-n…   16.8   
#> 15 Deer-Deer Proximity Rate, Captive (per day)     3.47   0.913 log-n…   53.5