Define starting compartment sizes for SIRS ODE Solver

Define starting compartment sizes for SIRS ODE Solver

Usage

initial_compartments(
  S_wild_prop = 1,
  I_wild_prop = 0,
  R_wild_prop = 0,
  S_captive_prop = 1,
  I_captive_prop = 0,
  R_captive_prop = 0,
  draws = NULL
)

Arguments

S_wild_prop

Size of susceptible compartment in wild population (proportion)

I_wild_prop

Size of infectious compartment in wild population (proportion)

R_wild_prop

Size of recovered compartment in wild population (proportion)

S_captive_prop

Size of susceptible compartment in captive population (proportion)

I_captive_prop

Size of infectious compartment in captive population (proportion)

R_captive_prop

Size of recovered compartment in captive population (proportion)

draws

$Number of iterations of simulation

I_wild_cumulative

Cumulative infected in wild segment, starting always at 0

I_captive_cumulative

Cumulative infected in captive segment, starting always at 0

Value

List of six vectors, repeating starting conditions specified by user. Length of vectors are determined by draws argument.

Details

Provides a vector of compartment sizes for wild and captive populations. This function pairs wild and captive deer populations so that transmission between populations can occur. Values (except for draws parameter) must be proportions, and must sum to one for each population.

Examples

initial_compartments(S_wild_prop = 0, draws = 5) #Starting conditions for a simulation focused only on captive deer
#> $S_wild
#> [1] 0 0 0 0 0
#> 
#> $I_wild
#> [1] 0 0 0 0 0
#> 
#> $R_wild
#> [1] 0 0 0 0 0
#> 
#> $I_wild_cumulative
#> [1] 0 0 0 0 0
#> 
#> $S_captive
#> [1] 1 1 1 1 1
#> 
#> $I_captive
#> [1] 0 0 0 0 0
#> 
#> $R_captive
#> [1] 0 0 0 0 0
#> 
#> $I_captive_cumulative
#> [1] 0 0 0 0 0
#> 
initial_compartments(draws = 5) #Starting conditions for a simulation of introduction and spread in both wild and captive deer.
#> $S_wild
#> [1] 1 1 1 1 1
#> 
#> $I_wild
#> [1] 0 0 0 0 0
#> 
#> $R_wild
#> [1] 0 0 0 0 0
#> 
#> $I_wild_cumulative
#> [1] 0 0 0 0 0
#> 
#> $S_captive
#> [1] 1 1 1 1 1
#> 
#> $I_captive
#> [1] 0 0 0 0 0
#> 
#> $R_captive
#> [1] 0 0 0 0 0
#> 
#> $I_captive_cumulative
#> [1] 0 0 0 0 0
#>