Simple function to generate SIRS model projections

simple_sirs is written to work alongside the deSolve::ode() function, and it will return an object with the proportion of individuals found in each of the SIR compartments at each of the specified time points.

Usage

simple_sirs(time, state, parameters)

Arguments

time

vector of values for which to generate the projection (length in days).

state

initial state for compartment populations. This should be a named vector for starting values of S_wild, I_wild, R_wild, S_captive, I_captive, R_captive.

parameters

list of parameters to generate projection. The parameters should include transmission parameters, immunity and recovery rates, and proportion of infected humans.

Value

when used with the deSolve::ode() function, it will return a dataframe with the proportion of individuals in each of the SIR compartments at each time point.

Examples

# prepare the input parameters:
example_inits <- c(S_wild = 1, I_wild = 0,
                   R_wild = 0, S_captive = 1,
                   I_captive = 0, R_captive = 0)

# set the time to run
example_times <-  seq(0, 500, by = 1)
# Set parameters of transmission, immunity, recovery

example_params <- c(alpha_immunity = 0.03,
                    beta_aero_ww = 0.01,
                    beta_aero_cw = 0.01,
                    beta_aero_cc = 0.02,
                    beta_aero_hw = 0.01,
                    beta_aero_hc = 0.2,
                    beta_dc_ww = 0.01,
                    beta_dc_cw = 0.01,
                    beta_dc_cc = 0.01,
                    phi_cw = 0,
                    phi_wc = 0,
                    gamma_recov = 0.01,
                    I_human = 0.05)

# run the ode function:

deSolve::ode(y = example_inits, times = example_times, parms = example_params, func = whitetailedSIRS::simple_sirs)
#>     time    S_wild       I_wild       R_wild S_captive  I_captive    R_captive
#> 1      0 1.0000000 0.0000000000 0.000000e+00 1.0000000 0.00000000 0.000000e+00
#> 2      1 0.9993937 0.0006033453 2.972120e-06 0.9898960 0.01005447 4.952954e-05
#> 3      2 0.9985719 0.0014152679 1.282810e-05 0.9795861 0.02021672 1.971848e-04
#> 4      3 0.9975291 0.0024394532 3.143763e-05 0.9690747 0.03048420 4.411150e-04
#> 5      4 0.9962606 0.0036789666 6.042078e-05 0.9583659 0.04085440 7.797231e-04
#> 6      5 0.9947604 0.0051376972 1.019176e-04 0.9474655 0.05132361 1.210903e-03
#> 7      6 0.9930238 0.0068185479 1.576558e-04 0.9363788 0.06188829 1.732952e-03
#> 8      7 0.9910463 0.0087243155 2.294167e-04 0.9251115 0.07254442 2.344106e-03
#> 9      8 0.9888235 0.0108575342 3.189569e-04 0.9136698 0.08328762 3.042611e-03
#> 10     9 0.9863516 0.0132203970 4.280010e-04 0.9020602 0.09411306 3.826719e-03
#> 11    10 0.9836270 0.0158147917 5.582446e-04 0.8902898 0.10501553 4.694687e-03
#> 12    11 0.9806464 0.0186422575 7.113503e-04 0.8783658 0.11598946 5.644766e-03
#> 13    12 0.9774071 0.0217039608 8.889445e-04 0.8662959 0.12702886 6.675204e-03
#> 14    13 0.9739067 0.0250006762 1.092615e-03 0.8540884 0.13812738 7.784242e-03
#> 15    14 0.9701433 0.0285327644 1.323910e-03 0.8417516 0.14927827 8.970104e-03
#> 16    15 0.9661155 0.0323001547 1.584329e-03 0.8292946 0.16047443 1.023100e-02
#> 17    16 0.9618223 0.0363023277 1.875330e-03 0.8167265 0.17170841 1.156513e-02
#> 18    17 0.9572634 0.0405382982 2.198316e-03 0.8040569 0.18297243 1.297065e-02
#> 19    18 0.9524388 0.0450066010 2.554638e-03 0.7912959 0.19425837 1.444571e-02
#> 20    19 0.9473491 0.0497052787 2.945591e-03 0.7784537 0.20555783 1.598844e-02
#> 21    20 0.9419957 0.0546318712 3.372409e-03 0.7655410 0.21686212 1.759692e-02
#> 22    21 0.9363803 0.0597834082 3.836263e-03 0.7525685 0.22816231 1.926922e-02
#> 23    22 0.9305053 0.0651564035 4.338258e-03 0.7395474 0.23944927 2.100335e-02
#> 24    23 0.9243737 0.0707468522 4.879429e-03 0.7264890 0.25071363 2.279733e-02
#> 25    24 0.9179890 0.0765502310 5.460738e-03 0.7134050 0.26194590 2.464910e-02
#> 26    25 0.9113554 0.0825615008 6.083072e-03 0.7003070 0.27313645 2.655660e-02
#> 27    26 0.9044776 0.0887751127 6.747240e-03 0.6872067 0.28427555 2.851773e-02
#> 28    27 0.8973610 0.0951850166 7.453969e-03 0.6741162 0.29535344 3.053033e-02
#> 29    28 0.8900114 0.1017846729 8.203903e-03 0.6610474 0.30636033 3.259225e-02
#> 30    29 0.8824353 0.1085670677 8.997600e-03 0.6480123 0.31728644 3.470126e-02
#> 31    30 0.8746397 0.1155247296 9.835530e-03 0.6350228 0.32812207 3.685515e-02
#> 32    31 0.8666322 0.1226497512 1.071807e-02 0.6220908 0.33885761 3.905164e-02
#> 33    32 0.8584207 0.1299338118 1.164552e-02 0.6092280 0.34948358 4.128843e-02
#> 34    33 0.8500137 0.1373682032 1.261807e-02 0.5964461 0.35999071 4.356323e-02
#> 35    34 0.8414203 0.1449438585 1.363582e-02 0.5837564 0.37036992 4.587367e-02
#> 36    35 0.8326498 0.1526513823 1.469878e-02 0.5711702 0.38061237 4.821741e-02
#> 37    36 0.8237120 0.1604810834 1.580688e-02 0.5586984 0.39070954 5.059208e-02
#> 38    37 0.8146171 0.1684230094 1.695992e-02 0.5463515 0.40065322 5.299528e-02
#> 39    38 0.8053754 0.1764669821 1.815764e-02 0.5341398 0.41043553 5.542463e-02
#> 40    39 0.7959977 0.1846026354 1.939966e-02 0.5220733 0.42004899 5.787772e-02
#> 41    40 0.7864950 0.1928194528 2.068554e-02 0.5101613 0.42948655 6.035215e-02
#> 42    41 0.7768785 0.2011068067 2.201472e-02 0.4984129 0.43874156 6.284552e-02
#> 43    42 0.7671594 0.2094539972 2.338657e-02 0.4868367 0.44780783 6.535545e-02
#> 44    43 0.7573494 0.2178502916 2.480035e-02 0.4754408 0.45667967 6.787956e-02
#> 45    44 0.7474598 0.2262849633 2.625527e-02 0.4642327 0.46535184 7.041549e-02
#> 46    45 0.7375022 0.2347473302 2.775042e-02 0.4532195 0.47381964 7.296089e-02
#> 47    46 0.7274884 0.2432267924 2.928485e-02 0.4424077 0.48207886 7.551346e-02
#> 48    47 0.7174296 0.2517128691 3.085751e-02 0.4318033 0.49012582 7.807090e-02
#> 49    48 0.7073375 0.2601952334 3.246728e-02 0.4214117 0.49795734 8.063098e-02
#> 50    49 0.6972233 0.2686637464 3.411297e-02 0.4112378 0.50557078 8.319146e-02
#> 51    50 0.6870982 0.2771084887 3.579335e-02 0.4012858 0.51296403 8.575019e-02
#> 52    51 0.6769731 0.2855197912 3.750709e-02 0.3915595 0.52013547 8.830503e-02
#> 53    52 0.6668589 0.2938882623 3.925285e-02 0.3820621 0.52708398 9.085391e-02
#> 54    53 0.6567660 0.3022048140 4.102921e-02 0.3727962 0.53380898 9.339481e-02
#> 55    54 0.6467046 0.3104606848 4.283470e-02 0.3637639 0.54031032 9.592576e-02
#> 56    55 0.6366847 0.3186474614 4.466783e-02 0.3549668 0.54658834 9.844485e-02
#> 57    56 0.6267158 0.3267570966 4.652706e-02 0.3464059 0.55264384 1.009502e-01
#> 58    57 0.6168073 0.3347819252 4.841082e-02 0.3380818 0.55847803 1.034401e-01
#> 59    58 0.6069678 0.3427146779 5.031752e-02 0.3299946 0.56409255 1.059128e-01
#> 60    59 0.5972060 0.3505484922 5.224552e-02 0.3221439 0.56948942 1.083667e-01
#> 61    60 0.5875299 0.3582769207 5.419321e-02 0.3145289 0.57467103 1.108001e-01
#> 62    61 0.5779471 0.3658939373 5.615892e-02 0.3071483 0.57964012 1.132116e-01
#> 63    62 0.5684651 0.3733939416 5.814100e-02 0.3000005 0.58439974 1.155997e-01
#> 64    63 0.5590905 0.3807717599 6.013778e-02 0.2930836 0.58895325 1.179632e-01
#> 65    64 0.5498297 0.3880226453 6.214761e-02 0.2863951 0.59330427 1.203006e-01
#> 66    65 0.5406889 0.3951422751 6.416883e-02 0.2799325 0.59745667 1.226108e-01
#> 67    66 0.5316735 0.4021267465 6.619978e-02 0.2736927 0.60141456 1.248927e-01
#> 68    67 0.5227886 0.4089725708 6.823883e-02 0.2676726 0.60518223 1.271452e-01
#> 69    68 0.5140390 0.4156766659 7.028436e-02 0.2618685 0.60876416 1.293673e-01
#> 70    69 0.5054289 0.4222363471 7.233477e-02 0.2562769 0.61216496 1.315582e-01
#> 71    70 0.4969622 0.4286493172 7.438848e-02 0.2508937 0.61538940 1.337169e-01
#> 72    71 0.4886424 0.4349136551 7.644394e-02 0.2457150 0.61844234 1.358427e-01
#> 73    72 0.4804726 0.4410278030 7.849962e-02 0.2407363 0.62132874 1.379349e-01
#> 74    73 0.4724554 0.4469905538 8.055403e-02 0.2359534 0.62405362 1.399930e-01
#> 75    74 0.4645933 0.4528010368 8.260571e-02 0.2313617 0.62662205 1.420163e-01
#> 76    75 0.4568881 0.4584587030 8.465324e-02 0.2269566 0.62903913 1.440043e-01
#> 77    76 0.4493415 0.4639633108 8.669523e-02 0.2227334 0.63130998 1.459567e-01
#> 78    77 0.4419547 0.4693149099 8.873035e-02 0.2186873 0.63343973 1.478730e-01
#> 79    78 0.4347289 0.4745138262 9.075728e-02 0.2148135 0.63543347 1.497530e-01
#> 80    79 0.4276646 0.4795606459 9.277477e-02 0.2111073 0.63729628 1.515964e-01
#> 81    80 0.4207622 0.4844562000 9.478161e-02 0.2075637 0.63903319 1.534031e-01
#> 82    81 0.4140218 0.4892015482 9.677662e-02 0.2041780 0.64064917 1.551728e-01
#> 83    82 0.4074434 0.4937979639 9.875868e-02 0.2009452 0.64214914 1.569057e-01
#> 84    83 0.4010264 0.4982469185 1.007267e-01 0.1978606 0.64353795 1.586015e-01
#> 85    84 0.3947703 0.5025500663 1.026797e-01 0.1949193 0.64482035 1.602604e-01
#> 86    85 0.3886742 0.5067092301 1.046166e-01 0.1921167 0.64600102 1.618823e-01
#> 87    86 0.3827371 0.5107263866 1.065366e-01 0.1894480 0.64708452 1.634675e-01
#> 88    87 0.3769577 0.5146036523 1.084387e-01 0.1869087 0.64807534 1.650160e-01
#> 89    88 0.3713347 0.5183432703 1.103220e-01 0.1844941 0.64897783 1.665281e-01
#> 90    89 0.3658665 0.5219475973 1.121859e-01 0.1821998 0.64979627 1.680040e-01
#> 91    90 0.3605513 0.5254190909 1.140296e-01 0.1800213 0.65053478 1.694439e-01
#> 92    91 0.3553874 0.5287602977 1.158523e-01 0.1779545 0.65119739 1.708481e-01
#> 93    92 0.3503727 0.5319738418 1.176535e-01 0.1759949 0.65178802 1.722170e-01
#> 94    93 0.3455052 0.5350624143 1.194324e-01 0.1741386 0.65231045 1.735510e-01
#> 95    94 0.3407827 0.5380287623 1.211886e-01 0.1723813 0.65276833 1.748503e-01
#> 96    95 0.3362028 0.5408756796 1.229215e-01 0.1707193 0.65316522 1.761155e-01
#> 97    96 0.3317633 0.5436059976 1.246307e-01 0.1691486 0.65350452 1.773468e-01
#> 98    97 0.3274618 0.5462225761 1.263156e-01 0.1676655 0.65378955 1.785449e-01
#> 99    98 0.3232957 0.5487282953 1.279760e-01 0.1662664 0.65402347 1.797101e-01
#> 100   99 0.3192624 0.5511260487 1.296115e-01 0.1649477 0.65420934 1.808429e-01
#> 101  100 0.3153595 0.5534187353 1.312218e-01 0.1637060 0.65435010 1.819439e-01
#> 102  101 0.3115842 0.5556092533 1.328065e-01 0.1625380 0.65444856 1.830135e-01
#> 103  102 0.3079340 0.5577004941 1.343655e-01 0.1614403 0.65450743 1.840522e-01
#> 104  103 0.3044061 0.5596953363 1.358985e-01 0.1604100 0.65452931 1.850607e-01
#> 105  104 0.3009979 0.5615966410 1.374054e-01 0.1594440 0.65451667 1.860394e-01
#> 106  105 0.2977067 0.5634072464 1.388861e-01 0.1585393 0.65447188 1.869888e-01
#> 107  106 0.2945296 0.5651299640 1.403404e-01 0.1576931 0.65439721 1.879096e-01
#> 108  107 0.2914642 0.5667675744 1.417683e-01 0.1569028 0.65429483 1.888024e-01
#> 109  108 0.2885075 0.5683228237 1.431696e-01 0.1561656 0.65416678 1.896676e-01
#> 110  109 0.2856570 0.5697984204 1.445445e-01 0.1554791 0.65401504 1.905058e-01
#> 111  110 0.2829100 0.5711970324 1.458929e-01 0.1548408 0.65384148 1.913177e-01
#> 112  111 0.2802638 0.5725212844 1.472149e-01 0.1542483 0.65364787 1.921038e-01
#> 113  112 0.2777157 0.5737737559 1.485105e-01 0.1536994 0.65343590 1.928647e-01
#> 114  113 0.2752632 0.5749569789 1.497798e-01 0.1531920 0.65320717 1.936008e-01
#> 115  114 0.2729037 0.5760734363 1.510229e-01 0.1527239 0.65296321 1.943129e-01
#> 116  115 0.2706345 0.5771255608 1.522399e-01 0.1522930 0.65270544 1.950015e-01
#> 117  116 0.2684532 0.5781157329 1.534311e-01 0.1518976 0.65243524 1.956671e-01
#> 118  117 0.2663573 0.5790462808 1.545964e-01 0.1515357 0.65215389 1.963104e-01
#> 119  118 0.2643443 0.5799194788 1.557363e-01 0.1512056 0.65186260 1.969318e-01
#> 120  119 0.2624117 0.5807375474 1.568507e-01 0.1509056 0.65156252 1.975319e-01
#> 121  120 0.2605573 0.5815026521 1.579400e-01 0.1506340 0.65125473 1.981113e-01
#> 122  121 0.2587786 0.5822169037 1.590045e-01 0.1503892 0.65094026 1.986705e-01
#> 123  122 0.2570734 0.5828823582 1.600442e-01 0.1501699 0.65062005 1.992101e-01
#> 124  123 0.2554394 0.5835010163 1.610595e-01 0.1499745 0.65029501 1.997305e-01
#> 125  124 0.2538744 0.5840748239 1.620507e-01 0.1498017 0.64996599 2.002323e-01
#> 126  125 0.2523762 0.5846056723 1.630181e-01 0.1496502 0.64963376 2.007160e-01
#> 127  126 0.2509427 0.5850953985 1.639619e-01 0.1495187 0.64929908 2.011822e-01
#> 128  127 0.2495718 0.5855457857 1.648824e-01 0.1494061 0.64896263 2.016312e-01
#> 129  128 0.2482615 0.5859585635 1.657799e-01 0.1493112 0.64862506 2.020637e-01
#> 130  129 0.2470097 0.5863354092 1.666549e-01 0.1492329 0.64828697 2.024801e-01
#> 131  130 0.2458146 0.5866779477 1.675075e-01 0.1491703 0.64794893 2.028808e-01
#> 132  131 0.2446742 0.5869877526 1.683381e-01 0.1491222 0.64761145 2.032664e-01
#> 133  132 0.2435866 0.5872663472 1.691470e-01 0.1490878 0.64727501 2.036372e-01
#> 134  133 0.2425501 0.5875152046 1.699347e-01 0.1490662 0.64694006 2.039938e-01
#> 135  134 0.2415629 0.5877357495 1.707014e-01 0.1490565 0.64660702 2.043365e-01
#> 136  135 0.2406232 0.5879293584 1.714475e-01 0.1490579 0.64627626 2.046659e-01
#> 137  136 0.2397294 0.5880973608 1.721732e-01 0.1490696 0.64594814 2.049822e-01
#> 138  137 0.2388798 0.5882410400 1.728791e-01 0.1490910 0.64562297 2.052860e-01
#> 139  138 0.2380729 0.5883616346 1.735654e-01 0.1491213 0.64530104 2.055777e-01
#> 140  139 0.2373071 0.5884603386 1.742325e-01 0.1491598 0.64498263 2.058575e-01
#> 141  140 0.2365809 0.5885383032 1.748808e-01 0.1492060 0.64466798 2.061260e-01
#> 142  141 0.2358928 0.5885966376 1.755106e-01 0.1492592 0.64435730 2.063834e-01
#> 143  142 0.2352414 0.5886364098 1.761222e-01 0.1493189 0.64405080 2.066303e-01
#> 144  143 0.2346253 0.5886586477 1.767161e-01 0.1493846 0.64374864 2.068668e-01
#> 145  144 0.2340431 0.5886643403 1.772926e-01 0.1494557 0.64345100 2.070933e-01
#> 146  145 0.2334936 0.5886544388 1.778520e-01 0.1495317 0.64315799 2.073103e-01
#> 147  146 0.2329755 0.5886298571 1.783946e-01 0.1496123 0.64286976 2.075180e-01
#> 148  147 0.2324875 0.5885914735 1.789210e-01 0.1496969 0.64258640 2.077167e-01
#> 149  148 0.2320285 0.5885401312 1.794313e-01 0.1497852 0.64230800 2.079068e-01
#> 150  149 0.2315973 0.5884766394 1.799260e-01 0.1498768 0.64203464 2.080886e-01
#> 151  150 0.2311928 0.5884017745 1.804054e-01 0.1499713 0.64176639 2.082623e-01
#> 152  151 0.2308139 0.5883162809 1.808698e-01 0.1500685 0.64150328 2.084283e-01
#> 153  152 0.2304595 0.5882208720 1.813197e-01 0.1501679 0.64124537 2.085868e-01
#> 154  153 0.2301286 0.5881162310 1.817552e-01 0.1502693 0.64099268 2.087380e-01
#> 155  154 0.2298202 0.5880030123 1.821768e-01 0.1503724 0.64074523 2.088824e-01
#> 156  155 0.2295334 0.5878818417 1.825848e-01 0.1504769 0.64050303 2.090201e-01
#> 157  156 0.2292672 0.5877533180 1.829794e-01 0.1505826 0.64026607 2.091513e-01
#> 158  157 0.2290208 0.5876180134 1.833612e-01 0.1506892 0.64003434 2.092764e-01
#> 159  158 0.2287933 0.5874764745 1.837302e-01 0.1507966 0.63980784 2.093955e-01
#> 160  159 0.2285838 0.5873292234 1.840870e-01 0.1509045 0.63958654 2.095089e-01
#> 161  160 0.2283915 0.5871767581 1.844317e-01 0.1510128 0.63937041 2.096168e-01
#> 162  161 0.2282157 0.5870195537 1.847647e-01 0.1511212 0.63915941 2.097194e-01
#> 163  162 0.2280556 0.5868580629 1.850863e-01 0.1512296 0.63895351 2.098169e-01
#> 164  163 0.2279105 0.5866927170 1.853968e-01 0.1513378 0.63875265 2.099095e-01
#> 165  164 0.2277796 0.5865239263 1.856965e-01 0.1514458 0.63855679 2.099974e-01
#> 166  165 0.2276623 0.5863520814 1.859856e-01 0.1515533 0.63836588 2.100809e-01
#> 167  166 0.2275580 0.5861775534 1.862645e-01 0.1516602 0.63817985 2.101600e-01
#> 168  167 0.2274659 0.5860006949 1.865334e-01 0.1517664 0.63799865 2.102349e-01
#> 169  168 0.2273856 0.5858218406 1.867926e-01 0.1518719 0.63782220 2.103059e-01
#> 170  169 0.2273164 0.5856413078 1.870423e-01 0.1519765 0.63765045 2.103731e-01
#> 171  170 0.2272577 0.5854593974 1.872829e-01 0.1520801 0.63748332 2.104366e-01
#> 172  171 0.2272090 0.5852763941 1.875146e-01 0.1521827 0.63732074 2.104966e-01
#> 173  172 0.2271698 0.5850925674 1.877377e-01 0.1522841 0.63716264 2.105532e-01
#> 174  173 0.2271395 0.5849081720 1.879523e-01 0.1523844 0.63700894 2.106067e-01
#> 175  174 0.2271178 0.5847234482 1.881588e-01 0.1524834 0.63685957 2.106571e-01
#> 176  175 0.2271041 0.5845386228 1.883573e-01 0.1525811 0.63671445 2.107045e-01
#> 177  176 0.2270979 0.5843539096 1.885482e-01 0.1526774 0.63657349 2.107491e-01
#> 178  177 0.2270989 0.5841695097 1.887316e-01 0.1527723 0.63643663 2.107911e-01
#> 179  178 0.2271066 0.5839856119 1.889078e-01 0.1528658 0.63630377 2.108304e-01
#> 180  179 0.2271207 0.5838023938 1.890769e-01 0.1529578 0.63617484 2.108673e-01
#> 181  180 0.2271407 0.5836200216 1.892393e-01 0.1530483 0.63604976 2.109019e-01
#> 182  181 0.2271663 0.5834386510 1.893950e-01 0.1531373 0.63592845 2.109343e-01
#> 183  182 0.2271972 0.5832584275 1.895444e-01 0.1532247 0.63581082 2.109645e-01
#> 184  183 0.2272329 0.5830794868 1.896876e-01 0.1533106 0.63569680 2.109926e-01
#> 185  184 0.2272732 0.5829019551 1.898248e-01 0.1533948 0.63558630 2.110189e-01
#> 186  185 0.2273178 0.5827259499 1.899562e-01 0.1534775 0.63547925 2.110433e-01
#> 187  186 0.2273664 0.5825515799 1.900820e-01 0.1535585 0.63537556 2.110659e-01
#> 188  187 0.2274187 0.5823789459 1.902024e-01 0.1536380 0.63527515 2.110869e-01
#> 189  188 0.2274744 0.5822081407 1.903175e-01 0.1537158 0.63517795 2.111062e-01
#> 190  189 0.2275332 0.5820392495 1.904275e-01 0.1537921 0.63508388 2.111241e-01
#> 191  190 0.2275950 0.5818723508 1.905327e-01 0.1538667 0.63499286 2.111405e-01
#> 192  191 0.2276594 0.5817075158 1.906331e-01 0.1539397 0.63490482 2.111555e-01
#> 193  192 0.2277263 0.5815448096 1.907289e-01 0.1540111 0.63481967 2.111693e-01
#> 194  193 0.2277954 0.5813842910 1.908203e-01 0.1540809 0.63473735 2.111818e-01
#> 195  194 0.2278666 0.5812260129 1.909074e-01 0.1541491 0.63465778 2.111931e-01
#> 196  195 0.2279396 0.5810700226 1.909904e-01 0.1542158 0.63458089 2.112034e-01
#> 197  196 0.2280142 0.5809163621 1.910694e-01 0.1542808 0.63450660 2.112126e-01
#> 198  197 0.2280903 0.5807650683 1.911446e-01 0.1543444 0.63443485 2.112208e-01
#> 199  198 0.2281677 0.5806161735 1.912161e-01 0.1544064 0.63436557 2.112280e-01
#> 200  199 0.2282463 0.5804697052 1.912840e-01 0.1544669 0.63429869 2.112344e-01
#> 201  200 0.2283258 0.5803256865 1.913485e-01 0.1545259 0.63423414 2.112400e-01
#> 202  201 0.2284062 0.5801841368 1.914097e-01 0.1545834 0.63417185 2.112447e-01
#> 203  202 0.2284873 0.5800450713 1.914677e-01 0.1546395 0.63411177 2.112487e-01
#> 204  203 0.2285689 0.5799085015 1.915226e-01 0.1546941 0.63405382 2.112520e-01
#> 205  204 0.2286511 0.5797744356 1.915745e-01 0.1547474 0.63399796 2.112547e-01
#> 206  205 0.2287335 0.5796428786 1.916236e-01 0.1547992 0.63394411 2.112567e-01
#> 207  206 0.2288162 0.5795138320 1.916700e-01 0.1548496 0.63389221 2.112582e-01
#> 208  207 0.2288990 0.5793872949 1.917137e-01 0.1548987 0.63384221 2.112591e-01
#> 209  208 0.2289818 0.5792632632 1.917549e-01 0.1549465 0.63379405 2.112595e-01
#> 210  209 0.2290645 0.5791417306 1.917937e-01 0.1549929 0.63374768 2.112594e-01
#> 211  210 0.2291471 0.5790226881 1.918302e-01 0.1550381 0.63370304 2.112589e-01
#> 212  211 0.2292295 0.5789061246 1.918644e-01 0.1550820 0.63366007 2.112579e-01
#> 213  212 0.2293115 0.5787920269 1.918965e-01 0.1551246 0.63361873 2.112566e-01
#> 214  213 0.2293931 0.5786803795 1.919265e-01 0.1551661 0.63357896 2.112549e-01
#> 215  214 0.2294743 0.5785711654 1.919545e-01 0.1552064 0.63354072 2.112529e-01
#> 216  215 0.2295550 0.5784643658 1.919807e-01 0.1552455 0.63350395 2.112506e-01
#> 217  216 0.2296350 0.5783599602 1.920050e-01 0.1552834 0.63346861 2.112480e-01
#> 218  217 0.2297145 0.5782579266 1.920276e-01 0.1553203 0.63343465 2.112451e-01
#> 219  218 0.2297932 0.5781582418 1.920485e-01 0.1553560 0.63340203 2.112420e-01
#> 220  219 0.2298713 0.5780608810 1.920679e-01 0.1553907 0.63337070 2.112386e-01
#> 221  220 0.2299485 0.5779658186 1.920857e-01 0.1554243 0.63334062 2.112351e-01
#> 222  221 0.2300249 0.5778730277 1.921021e-01 0.1554569 0.63331175 2.112313e-01
#> 223  222 0.2301005 0.5777824804 1.921170e-01 0.1554885 0.63328405 2.112274e-01
#> 224  223 0.2301751 0.5776941481 1.921307e-01 0.1555192 0.63325747 2.112234e-01
#> 225  224 0.2302489 0.5776080012 1.921431e-01 0.1555488 0.63323199 2.112192e-01
#> 226  225 0.2303217 0.5775240094 1.921543e-01 0.1555776 0.63320756 2.112149e-01
#> 227  226 0.2303935 0.5774421419 1.921643e-01 0.1556054 0.63318414 2.112104e-01
#> 228  227 0.2304643 0.5773623671 1.921733e-01 0.1556324 0.63316171 2.112059e-01
#> 229  228 0.2305341 0.5772846531 1.921812e-01 0.1556585 0.63314022 2.112013e-01
#> 230  229 0.2306029 0.5772089673 1.921881e-01 0.1556837 0.63311965 2.111967e-01
#> 231  230 0.2306706 0.5771352769 1.921941e-01 0.1557081 0.63309996 2.111919e-01
#> 232  231 0.2307372 0.5770635489 1.921992e-01 0.1557317 0.63308112 2.111871e-01
#> 233  232 0.2308028 0.5769937496 1.922035e-01 0.1557546 0.63306309 2.111823e-01
#> 234  233 0.2308673 0.5769258456 1.922069e-01 0.1557767 0.63304586 2.111775e-01
#> 235  234 0.2309306 0.5768598029 1.922096e-01 0.1557980 0.63302939 2.111726e-01
#> 236  235 0.2309929 0.5767955876 1.922115e-01 0.1558186 0.63301365 2.111677e-01
#> 237  236 0.2310540 0.5767331657 1.922128e-01 0.1558386 0.63299862 2.111628e-01
#> 238  237 0.2311141 0.5766725031 1.922134e-01 0.1558578 0.63298427 2.111579e-01
#> 239  238 0.2311730 0.5766135659 1.922135e-01 0.1558764 0.63297057 2.111530e-01
#> 240  239 0.2312308 0.5765563200 1.922129e-01 0.1558944 0.63295751 2.111481e-01
#> 241  240 0.2312874 0.5765007315 1.922118e-01 0.1559117 0.63294505 2.111433e-01
#> 242  241 0.2313430 0.5764467666 1.922102e-01 0.1559284 0.63293317 2.111384e-01
#> 243  242 0.2313975 0.5763943917 1.922082e-01 0.1559445 0.63292186 2.111336e-01
#> 244  243 0.2314508 0.5763435732 1.922056e-01 0.1559601 0.63291108 2.111289e-01
#> 245  244 0.2315030 0.5762942778 1.922027e-01 0.1559750 0.63290082 2.111241e-01
#> 246  245 0.2315542 0.5762464726 1.921994e-01 0.1559895 0.63289106 2.111194e-01
#> 247  246 0.2316042 0.5762001247 1.921957e-01 0.1560034 0.63288178 2.111148e-01
#> 248  247 0.2316532 0.5761552016 1.921916e-01 0.1560169 0.63287296 2.111102e-01
#> 249  248 0.2317010 0.5761116710 1.921873e-01 0.1560298 0.63286458 2.111056e-01
#> 250  249 0.2317479 0.5760695010 1.921826e-01 0.1560422 0.63285663 2.111011e-01
#> 251  250 0.2317936 0.5760286599 1.921777e-01 0.1560542 0.63284908 2.110967e-01
#> 252  251 0.2318383 0.5759891167 1.921726e-01 0.1560658 0.63284192 2.110923e-01
#> 253  252 0.2318820 0.5759508402 1.921672e-01 0.1560769 0.63283513 2.110880e-01
#> 254  253 0.2319246 0.5759138001 1.921616e-01 0.1560875 0.63282871 2.110837e-01
#> 255  254 0.2319663 0.5758779662 1.921557e-01 0.1560978 0.63282263 2.110796e-01
#> 256  255 0.2320069 0.5758433087 1.921498e-01 0.1561077 0.63281687 2.110754e-01
#> 257  256 0.2320466 0.5758097983 1.921436e-01 0.1561172 0.63281144 2.110714e-01
#> 258  257 0.2320852 0.5757774061 1.921374e-01 0.1561263 0.63280630 2.110674e-01
#> 259  258 0.2321230 0.5757461036 1.921309e-01 0.1561351 0.63280146 2.110634e-01
#> 260  259 0.2321597 0.5757158628 1.921244e-01 0.1561435 0.63279689 2.110596e-01
#> 261  260 0.2321955 0.5756866559 1.921178e-01 0.1561516 0.63279259 2.110558e-01
#> 262  261 0.2322305 0.5756584559 1.921111e-01 0.1561594 0.63278854 2.110521e-01
#> 263  262 0.2322645 0.5756312360 1.921043e-01 0.1561668 0.63278473 2.110485e-01
#> 264  263 0.2322976 0.5756049700 1.920975e-01 0.1561740 0.63278115 2.110449e-01
#> 265  264 0.2323298 0.5755796320 1.920906e-01 0.1561808 0.63277780 2.110414e-01
#> 266  265 0.2323612 0.5755551966 1.920836e-01 0.1561874 0.63277465 2.110380e-01
#> 267  266 0.2323917 0.5755316389 1.920767e-01 0.1561937 0.63277171 2.110346e-01
#> 268  267 0.2324214 0.5755089346 1.920697e-01 0.1561997 0.63276896 2.110313e-01
#> 269  268 0.2324503 0.5754870596 1.920627e-01 0.1562055 0.63276640 2.110281e-01
#> 270  269 0.2324784 0.5754659903 1.920556e-01 0.1562110 0.63276401 2.110250e-01
#> 271  270 0.2325057 0.5754457039 1.920486e-01 0.1562163 0.63276179 2.110219e-01
#> 272  271 0.2325322 0.5754261775 1.920416e-01 0.1562214 0.63275972 2.110189e-01
#> 273  272 0.2325579 0.5754073892 1.920347e-01 0.1562262 0.63275781 2.110159e-01
#> 274  273 0.2325830 0.5753893171 1.920277e-01 0.1562309 0.63275604 2.110131e-01
#> 275  274 0.2326073 0.5753719402 1.920208e-01 0.1562353 0.63275441 2.110103e-01
#> 276  275 0.2326309 0.5753552375 1.920139e-01 0.1562395 0.63275291 2.110075e-01
#> 277  276 0.2326537 0.5753391888 1.920071e-01 0.1562436 0.63275153 2.110049e-01
#> 278  277 0.2326760 0.5753237743 1.920003e-01 0.1562474 0.63275027 2.110023e-01
#> 279  278 0.2326975 0.5753089743 1.919935e-01 0.1562511 0.63274912 2.109998e-01
#> 280  279 0.2327184 0.5752947700 1.919868e-01 0.1562546 0.63274808 2.109973e-01
#> 281  280 0.2327386 0.5752811428 1.919802e-01 0.1562580 0.63274714 2.109949e-01
#> 282  281 0.2327583 0.5752680745 1.919736e-01 0.1562612 0.63274630 2.109925e-01
#> 283  282 0.2327773 0.5752555475 1.919672e-01 0.1562642 0.63274554 2.109903e-01
#> 284  283 0.2327957 0.5752435443 1.919607e-01 0.1562671 0.63274487 2.109880e-01
#> 285  284 0.2328136 0.5752320483 1.919544e-01 0.1562699 0.63274428 2.109859e-01
#> 286  285 0.2328308 0.5752210428 1.919481e-01 0.1562725 0.63274377 2.109838e-01
#> 287  286 0.2328476 0.5752105118 1.919419e-01 0.1562750 0.63274333 2.109817e-01
#> 288  287 0.2328638 0.5752004396 1.919358e-01 0.1562773 0.63274296 2.109797e-01
#> 289  288 0.2328794 0.5751908110 1.919298e-01 0.1562796 0.63274265 2.109778e-01
#> 290  289 0.2328946 0.5751816112 1.919238e-01 0.1562817 0.63274240 2.109759e-01
#> 291  290 0.2329092 0.5751728255 1.919180e-01 0.1562837 0.63274221 2.109741e-01
#> 292  291 0.2329233 0.5751644399 1.919122e-01 0.1562856 0.63274207 2.109723e-01
#> 293  292 0.2329370 0.5751564406 1.919065e-01 0.1562874 0.63274199 2.109706e-01
#> 294  293 0.2329502 0.5751488142 1.919010e-01 0.1562891 0.63274195 2.109689e-01
#> 295  294 0.2329630 0.5751415477 1.918955e-01 0.1562908 0.63274195 2.109673e-01
#> 296  295 0.2329753 0.5751346284 1.918901e-01 0.1562923 0.63274200 2.109657e-01
#> 297  296 0.2329872 0.5751280439 1.918847e-01 0.1562937 0.63274208 2.109642e-01
#> 298  297 0.2329987 0.5751217824 1.918795e-01 0.1562951 0.63274220 2.109627e-01
#> 299  298 0.2330098 0.5751158320 1.918744e-01 0.1562964 0.63274236 2.109613e-01
#> 300  299 0.2330204 0.5751101815 1.918694e-01 0.1562976 0.63274254 2.109599e-01
#> 301  300 0.2330307 0.5751048199 1.918645e-01 0.1562987 0.63274276 2.109585e-01
#> 302  301 0.2330407 0.5750997364 1.918596e-01 0.1562998 0.63274300 2.109572e-01
#> 303  302 0.2330502 0.5750949207 1.918549e-01 0.1563008 0.63274327 2.109560e-01
#> 304  303 0.2330594 0.5750903627 1.918502e-01 0.1563017 0.63274356 2.109547e-01
#> 305  304 0.2330683 0.5750860525 1.918457e-01 0.1563026 0.63274387 2.109535e-01
#> 306  305 0.2330768 0.5750819807 1.918412e-01 0.1563034 0.63274420 2.109524e-01
#> 307  306 0.2330850 0.5750781380 1.918368e-01 0.1563042 0.63274455 2.109513e-01
#> 308  307 0.2330929 0.5750745156 1.918326e-01 0.1563049 0.63274491 2.109502e-01
#> 309  308 0.2331005 0.5750711046 1.918284e-01 0.1563055 0.63274529 2.109492e-01
#> 310  309 0.2331078 0.5750678968 1.918243e-01 0.1563062 0.63274568 2.109482e-01
#> 311  310 0.2331148 0.5750648840 1.918203e-01 0.1563067 0.63274609 2.109472e-01
#> 312  311 0.2331216 0.5750620582 1.918164e-01 0.1563072 0.63274651 2.109463e-01
#> 313  312 0.2331281 0.5750594119 1.918125e-01 0.1563077 0.63274693 2.109453e-01
#> 314  313 0.2331343 0.5750569376 1.918088e-01 0.1563082 0.63274736 2.109445e-01
#> 315  314 0.2331402 0.5750546281 1.918051e-01 0.1563086 0.63274780 2.109436e-01
#> 316  315 0.2331459 0.5750524766 1.918016e-01 0.1563089 0.63274825 2.109428e-01
#> 317  316 0.2331514 0.5750504762 1.917981e-01 0.1563093 0.63274870 2.109420e-01
#> 318  317 0.2331567 0.5750486206 1.917947e-01 0.1563096 0.63274916 2.109413e-01
#> 319  318 0.2331617 0.5750469034 1.917914e-01 0.1563099 0.63274962 2.109405e-01
#> 320  319 0.2331665 0.5750453185 1.917882e-01 0.1563101 0.63275009 2.109398e-01
#> 321  320 0.2331711 0.5750438601 1.917850e-01 0.1563103 0.63275055 2.109391e-01
#> 322  321 0.2331755 0.5750425225 1.917819e-01 0.1563105 0.63275102 2.109385e-01
#> 323  322 0.2331798 0.5750413003 1.917789e-01 0.1563107 0.63275149 2.109379e-01
#> 324  323 0.2331838 0.5750401880 1.917760e-01 0.1563108 0.63275195 2.109372e-01
#> 325  324 0.2331876 0.5750391807 1.917732e-01 0.1563109 0.63275242 2.109367e-01
#> 326  325 0.2331913 0.5750382734 1.917704e-01 0.1563110 0.63275289 2.109361e-01
#> 327  326 0.2331948 0.5750374614 1.917677e-01 0.1563111 0.63275335 2.109355e-01
#> 328  327 0.2331982 0.5750367400 1.917651e-01 0.1563112 0.63275381 2.109350e-01
#> 329  328 0.2332013 0.5750361049 1.917625e-01 0.1563112 0.63275427 2.109345e-01
#> 330  329 0.2332044 0.5750355517 1.917601e-01 0.1563112 0.63275473 2.109340e-01
#> 331  330 0.2332073 0.5750350764 1.917577e-01 0.1563112 0.63275518 2.109336e-01
#> 332  331 0.2332100 0.5750346749 1.917553e-01 0.1563112 0.63275563 2.109331e-01
#> 333  332 0.2332126 0.5750343436 1.917530e-01 0.1563112 0.63275608 2.109327e-01
#> 334  333 0.2332151 0.5750340787 1.917508e-01 0.1563112 0.63275652 2.109323e-01
#> 335  334 0.2332175 0.5750338766 1.917487e-01 0.1563111 0.63275696 2.109319e-01
#> 336  335 0.2332197 0.5750337341 1.917466e-01 0.1563111 0.63275739 2.109315e-01
#> 337  336 0.2332218 0.5750336477 1.917445e-01 0.1563110 0.63275782 2.109311e-01
#> 338  337 0.2332238 0.5750336145 1.917426e-01 0.1563110 0.63275824 2.109308e-01
#> 339  338 0.2332257 0.5750336313 1.917407e-01 0.1563109 0.63275866 2.109305e-01
#> 340  339 0.2332275 0.5750336952 1.917388e-01 0.1563108 0.63275907 2.109301e-01
#> 341  340 0.2332292 0.5750338035 1.917370e-01 0.1563107 0.63275948 2.109298e-01
#> 342  341 0.2332308 0.5750339535 1.917353e-01 0.1563106 0.63275988 2.109295e-01
#> 343  342 0.2332323 0.5750341427 1.917336e-01 0.1563105 0.63276027 2.109293e-01
#> 344  343 0.2332337 0.5750343685 1.917319e-01 0.1563103 0.63276066 2.109290e-01
#> 345  344 0.2332350 0.5750346286 1.917303e-01 0.1563102 0.63276104 2.109287e-01
#> 346  345 0.2332363 0.5750349207 1.917288e-01 0.1563101 0.63276141 2.109285e-01
#> 347  346 0.2332374 0.5750352427 1.917273e-01 0.1563100 0.63276178 2.109283e-01
#> 348  347 0.2332385 0.5750355925 1.917259e-01 0.1563098 0.63276215 2.109280e-01
#> 349  348 0.2332396 0.5750359682 1.917245e-01 0.1563097 0.63276250 2.109278e-01
#> 350  349 0.2332405 0.5750363677 1.917231e-01 0.1563095 0.63276285 2.109276e-01
#> 351  350 0.2332414 0.5750367893 1.917218e-01 0.1563094 0.63276319 2.109274e-01
#> 352  351 0.2332422 0.5750372312 1.917205e-01 0.1563092 0.63276353 2.109272e-01
#> 353  352 0.2332430 0.5750376919 1.917193e-01 0.1563091 0.63276386 2.109271e-01
#> 354  353 0.2332437 0.5750381696 1.917181e-01 0.1563089 0.63276418 2.109269e-01
#> 355  354 0.2332443 0.5750386630 1.917170e-01 0.1563088 0.63276450 2.109267e-01
#> 356  355 0.2332449 0.5750391704 1.917159e-01 0.1563086 0.63276481 2.109266e-01
#> 357  356 0.2332455 0.5750396907 1.917148e-01 0.1563085 0.63276512 2.109264e-01
#> 358  357 0.2332460 0.5750402224 1.917138e-01 0.1563083 0.63276541 2.109263e-01
#> 359  358 0.2332464 0.5750407644 1.917128e-01 0.1563081 0.63276571 2.109262e-01
#> 360  359 0.2332468 0.5750413153 1.917118e-01 0.1563080 0.63276599 2.109260e-01
#> 361  360 0.2332472 0.5750418742 1.917109e-01 0.1563078 0.63276627 2.109259e-01
#> 362  361 0.2332475 0.5750424399 1.917100e-01 0.1563077 0.63276654 2.109258e-01
#> 363  362 0.2332478 0.5750430114 1.917092e-01 0.1563075 0.63276681 2.109257e-01
#> 364  363 0.2332481 0.5750435878 1.917083e-01 0.1563073 0.63276707 2.109256e-01
#> 365  364 0.2332483 0.5750441680 1.917075e-01 0.1563072 0.63276732 2.109255e-01
#> 366  365 0.2332485 0.5750447514 1.917068e-01 0.1563070 0.63276757 2.109254e-01
#> 367  366 0.2332486 0.5750453370 1.917060e-01 0.1563069 0.63276781 2.109253e-01
#> 368  367 0.2332488 0.5750459240 1.917053e-01 0.1563067 0.63276805 2.109252e-01
#> 369  368 0.2332489 0.5750465118 1.917046e-01 0.1563065 0.63276828 2.109252e-01
#> 370  369 0.2332489 0.5750470996 1.917040e-01 0.1563064 0.63276851 2.109251e-01
#> 371  370 0.2332490 0.5750476868 1.917033e-01 0.1563062 0.63276873 2.109250e-01
#> 372  371 0.2332490 0.5750482728 1.917027e-01 0.1563061 0.63276894 2.109250e-01
#> 373  372 0.2332490 0.5750488570 1.917021e-01 0.1563059 0.63276915 2.109249e-01
#> 374  373 0.2332490 0.5750494389 1.917016e-01 0.1563058 0.63276935 2.109249e-01
#> 375  374 0.2332490 0.5750500179 1.917010e-01 0.1563056 0.63276955 2.109248e-01
#> 376  375 0.2332489 0.5750505936 1.917005e-01 0.1563055 0.63276974 2.109248e-01
#> 377  376 0.2332488 0.5750511655 1.917000e-01 0.1563053 0.63276993 2.109247e-01
#> 378  377 0.2332487 0.5750517332 1.916995e-01 0.1563052 0.63277011 2.109247e-01
#> 379  378 0.2332486 0.5750522964 1.916991e-01 0.1563051 0.63277029 2.109246e-01
#> 380  379 0.2332485 0.5750528546 1.916986e-01 0.1563049 0.63277046 2.109246e-01
#> 381  380 0.2332484 0.5750534075 1.916982e-01 0.1563048 0.63277063 2.109246e-01
#> 382  381 0.2332482 0.5750539550 1.916978e-01 0.1563047 0.63277080 2.109245e-01
#> 383  382 0.2332481 0.5750544965 1.916974e-01 0.1563045 0.63277096 2.109245e-01
#> 384  383 0.2332479 0.5750550320 1.916970e-01 0.1563044 0.63277111 2.109245e-01
#> 385  384 0.2332477 0.5750555611 1.916967e-01 0.1563043 0.63277126 2.109245e-01
#> 386  385 0.2332476 0.5750560836 1.916964e-01 0.1563041 0.63277141 2.109245e-01
#> 387  386 0.2332474 0.5750565994 1.916960e-01 0.1563040 0.63277155 2.109244e-01
#> 388  387 0.2332472 0.5750571083 1.916957e-01 0.1563039 0.63277169 2.109244e-01
#> 389  388 0.2332470 0.5750576100 1.916954e-01 0.1563038 0.63277182 2.109244e-01
#> 390  389 0.2332467 0.5750581045 1.916952e-01 0.1563037 0.63277195 2.109244e-01
#> 391  390 0.2332465 0.5750585916 1.916949e-01 0.1563035 0.63277208 2.109244e-01
#> 392  391 0.2332463 0.5750590713 1.916946e-01 0.1563034 0.63277220 2.109244e-01
#> 393  392 0.2332461 0.5750595433 1.916944e-01 0.1563033 0.63277232 2.109244e-01
#> 394  393 0.2332458 0.5750600077 1.916942e-01 0.1563032 0.63277243 2.109244e-01
#> 395  394 0.2332456 0.5750604643 1.916939e-01 0.1563031 0.63277254 2.109243e-01
#> 396  395 0.2332454 0.5750609132 1.916937e-01 0.1563030 0.63277265 2.109243e-01
#> 397  396 0.2332451 0.5750613542 1.916935e-01 0.1563029 0.63277276 2.109243e-01
#> 398  397 0.2332449 0.5750617873 1.916933e-01 0.1563028 0.63277286 2.109243e-01
#> 399  398 0.2332446 0.5750622125 1.916932e-01 0.1563027 0.63277296 2.109243e-01
#> 400  399 0.2332444 0.5750626298 1.916930e-01 0.1563026 0.63277305 2.109243e-01
#> 401  400 0.2332441 0.5750630392 1.916928e-01 0.1563025 0.63277315 2.109243e-01
#> 402  401 0.2332439 0.5750634407 1.916927e-01 0.1563024 0.63277323 2.109243e-01
#> 403  402 0.2332436 0.5750638342 1.916925e-01 0.1563023 0.63277332 2.109243e-01
#> 404  403 0.2332434 0.5750642199 1.916924e-01 0.1563023 0.63277340 2.109243e-01
#> 405  404 0.2332431 0.5750645977 1.916923e-01 0.1563022 0.63277349 2.109243e-01
#> 406  405 0.2332429 0.5750649677 1.916922e-01 0.1563021 0.63277356 2.109244e-01
#> 407  406 0.2332426 0.5750653300 1.916921e-01 0.1563020 0.63277364 2.109244e-01
#> 408  407 0.2332424 0.5750656845 1.916920e-01 0.1563019 0.63277371 2.109244e-01
#> 409  408 0.2332421 0.5750660313 1.916919e-01 0.1563018 0.63277378 2.109244e-01
#> 410  409 0.2332419 0.5750663706 1.916918e-01 0.1563018 0.63277385 2.109244e-01
#> 411  410 0.2332416 0.5750667022 1.916917e-01 0.1563017 0.63277392 2.109244e-01
#> 412  411 0.2332414 0.5750670264 1.916916e-01 0.1563016 0.63277398 2.109244e-01
#> 413  412 0.2332411 0.5750673432 1.916915e-01 0.1563016 0.63277404 2.109244e-01
#> 414  413 0.2332409 0.5750676527 1.916915e-01 0.1563015 0.63277410 2.109244e-01
#> 415  414 0.2332407 0.5750679549 1.916914e-01 0.1563014 0.63277416 2.109244e-01
#> 416  415 0.2332404 0.5750682500 1.916913e-01 0.1563014 0.63277422 2.109244e-01
#> 417  416 0.2332402 0.5750685379 1.916913e-01 0.1563013 0.63277427 2.109244e-01
#> 418  417 0.2332400 0.5750688189 1.916912e-01 0.1563012 0.63277432 2.109244e-01
#> 419  418 0.2332397 0.5750690930 1.916912e-01 0.1563012 0.63277437 2.109245e-01
#> 420  419 0.2332395 0.5750693603 1.916911e-01 0.1563011 0.63277442 2.109245e-01
#> 421  420 0.2332393 0.5750696209 1.916911e-01 0.1563011 0.63277446 2.109245e-01
#> 422  421 0.2332391 0.5750698749 1.916911e-01 0.1563010 0.63277451 2.109245e-01
#> 423  422 0.2332389 0.5750701224 1.916910e-01 0.1563009 0.63277455 2.109245e-01
#> 424  423 0.2332386 0.5750703635 1.916910e-01 0.1563009 0.63277459 2.109245e-01
#> 425  424 0.2332384 0.5750705983 1.916910e-01 0.1563008 0.63277463 2.109245e-01
#> 426  425 0.2332382 0.5750708269 1.916910e-01 0.1563008 0.63277467 2.109245e-01
#> 427  426 0.2332380 0.5750710494 1.916909e-01 0.1563008 0.63277471 2.109245e-01
#> 428  427 0.2332378 0.5750712660 1.916909e-01 0.1563007 0.63277474 2.109246e-01
#> 429  428 0.2332376 0.5750714766 1.916909e-01 0.1563007 0.63277478 2.109246e-01
#> 430  429 0.2332374 0.5750716815 1.916909e-01 0.1563006 0.63277481 2.109246e-01
#> 431  430 0.2332372 0.5750718807 1.916909e-01 0.1563006 0.63277484 2.109246e-01
#> 432  431 0.2332370 0.5750720744 1.916909e-01 0.1563005 0.63277487 2.109246e-01
#> 433  432 0.2332369 0.5750722626 1.916909e-01 0.1563005 0.63277490 2.109246e-01
#> 434  433 0.2332367 0.5750724455 1.916909e-01 0.1563005 0.63277493 2.109246e-01
#> 435  434 0.2332365 0.5750726231 1.916909e-01 0.1563004 0.63277495 2.109246e-01
#> 436  435 0.2332363 0.5750727956 1.916909e-01 0.1563004 0.63277498 2.109246e-01
#> 437  436 0.2332362 0.5750729631 1.916909e-01 0.1563003 0.63277500 2.109246e-01
#> 438  437 0.2332360 0.5750731256 1.916909e-01 0.1563003 0.63277503 2.109247e-01
#> 439  438 0.2332358 0.5750732833 1.916909e-01 0.1563003 0.63277505 2.109247e-01
#> 440  439 0.2332357 0.5750734364 1.916909e-01 0.1563002 0.63277507 2.109247e-01
#> 441  440 0.2332355 0.5750735848 1.916909e-01 0.1563002 0.63277509 2.109247e-01
#> 442  441 0.2332354 0.5750737287 1.916909e-01 0.1563002 0.63277511 2.109247e-01
#> 443  442 0.2332352 0.5750738682 1.916909e-01 0.1563002 0.63277513 2.109247e-01
#> 444  443 0.2332351 0.5750740034 1.916909e-01 0.1563001 0.63277515 2.109247e-01
#> 445  444 0.2332349 0.5750741343 1.916909e-01 0.1563001 0.63277517 2.109247e-01
#> 446  445 0.2332348 0.5750742612 1.916910e-01 0.1563001 0.63277518 2.109247e-01
#> 447  446 0.2332346 0.5750743841 1.916910e-01 0.1563001 0.63277520 2.109248e-01
#> 448  447 0.2332345 0.5750745030 1.916910e-01 0.1563000 0.63277521 2.109248e-01
#> 449  448 0.2332344 0.5750746181 1.916910e-01 0.1563000 0.63277523 2.109248e-01
#> 450  449 0.2332343 0.5750747295 1.916910e-01 0.1563000 0.63277524 2.109248e-01
#> 451  450 0.2332341 0.5750748373 1.916910e-01 0.1563000 0.63277526 2.109248e-01
#> 452  451 0.2332340 0.5750749415 1.916911e-01 0.1562999 0.63277527 2.109248e-01
#> 453  452 0.2332339 0.5750750422 1.916911e-01 0.1562999 0.63277528 2.109248e-01
#> 454  453 0.2332338 0.5750751396 1.916911e-01 0.1562999 0.63277529 2.109248e-01
#> 455  454 0.2332337 0.5750752337 1.916911e-01 0.1562999 0.63277530 2.109248e-01
#> 456  455 0.2332335 0.5750753246 1.916911e-01 0.1562999 0.63277531 2.109248e-01
#> 457  456 0.2332334 0.5750754124 1.916911e-01 0.1562998 0.63277532 2.109248e-01
#> 458  457 0.2332333 0.5750754972 1.916912e-01 0.1562998 0.63277533 2.109248e-01
#> 459  458 0.2332332 0.5750755790 1.916912e-01 0.1562998 0.63277534 2.109249e-01
#> 460  459 0.2332331 0.5750756580 1.916912e-01 0.1562998 0.63277535 2.109249e-01
#> 461  460 0.2332330 0.5750757341 1.916912e-01 0.1562998 0.63277536 2.109249e-01
#> 462  461 0.2332329 0.5750758076 1.916912e-01 0.1562998 0.63277536 2.109249e-01
#> 463  462 0.2332329 0.5750758784 1.916913e-01 0.1562997 0.63277537 2.109249e-01
#> 464  463 0.2332328 0.5750759466 1.916913e-01 0.1562997 0.63277538 2.109249e-01
#> 465  464 0.2332327 0.5750760124 1.916913e-01 0.1562997 0.63277538 2.109249e-01
#> 466  465 0.2332326 0.5750760757 1.916913e-01 0.1562997 0.63277539 2.109249e-01
#> 467  466 0.2332325 0.5750761367 1.916914e-01 0.1562997 0.63277540 2.109249e-01
#> 468  467 0.2332324 0.5750761954 1.916914e-01 0.1562997 0.63277540 2.109249e-01
#> 469  468 0.2332324 0.5750762518 1.916914e-01 0.1562997 0.63277541 2.109249e-01
#> 470  469 0.2332323 0.5750763062 1.916914e-01 0.1562997 0.63277541 2.109249e-01
#> 471  470 0.2332322 0.5750763584 1.916914e-01 0.1562996 0.63277542 2.109249e-01
#> 472  471 0.2332321 0.5750764086 1.916915e-01 0.1562996 0.63277542 2.109249e-01
#> 473  472 0.2332321 0.5750764569 1.916915e-01 0.1562996 0.63277542 2.109249e-01
#> 474  473 0.2332320 0.5750765033 1.916915e-01 0.1562996 0.63277543 2.109250e-01
#> 475  474 0.2332319 0.5750765478 1.916915e-01 0.1562996 0.63277543 2.109250e-01
#> 476  475 0.2332319 0.5750765905 1.916915e-01 0.1562996 0.63277543 2.109250e-01
#> 477  476 0.2332318 0.5750766315 1.916916e-01 0.1562996 0.63277544 2.109250e-01
#> 478  477 0.2332318 0.5750766708 1.916916e-01 0.1562996 0.63277544 2.109250e-01
#> 479  478 0.2332317 0.5750767085 1.916916e-01 0.1562996 0.63277544 2.109250e-01
#> 480  479 0.2332316 0.5750767446 1.916916e-01 0.1562996 0.63277544 2.109250e-01
#> 481  480 0.2332316 0.5750767792 1.916916e-01 0.1562996 0.63277545 2.109250e-01
#> 482  481 0.2332315 0.5750768124 1.916917e-01 0.1562996 0.63277545 2.109250e-01
#> 483  482 0.2332315 0.5750768441 1.916917e-01 0.1562996 0.63277545 2.109250e-01
#> 484  483 0.2332314 0.5750768744 1.916917e-01 0.1562995 0.63277545 2.109250e-01
#> 485  484 0.2332314 0.5750769034 1.916917e-01 0.1562995 0.63277545 2.109250e-01
#> 486  485 0.2332313 0.5750769312 1.916917e-01 0.1562995 0.63277545 2.109250e-01
#> 487  486 0.2332313 0.5750769576 1.916917e-01 0.1562995 0.63277545 2.109250e-01
#> 488  487 0.2332313 0.5750769829 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 489  488 0.2332312 0.5750770070 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 490  489 0.2332312 0.5750770301 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 491  490 0.2332311 0.5750770520 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 492  491 0.2332311 0.5750770729 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 493  492 0.2332311 0.5750770928 1.916918e-01 0.1562995 0.63277546 2.109250e-01
#> 494  493 0.2332310 0.5750771117 1.916919e-01 0.1562995 0.63277546 2.109250e-01
#> 495  494 0.2332310 0.5750771297 1.916919e-01 0.1562995 0.63277546 2.109250e-01
#> 496  495 0.2332310 0.5750771467 1.916919e-01 0.1562995 0.63277546 2.109250e-01
#> 497  496 0.2332309 0.5750771630 1.916919e-01 0.1562995 0.63277546 2.109251e-01
#> 498  497 0.2332309 0.5750771783 1.916919e-01 0.1562995 0.63277546 2.109251e-01
#> 499  498 0.2332309 0.5750771929 1.916919e-01 0.1562995 0.63277546 2.109251e-01
#> 500  499 0.2332308 0.5750772067 1.916919e-01 0.1562995 0.63277546 2.109251e-01
#> 501  500 0.2332308 0.5750772198 1.916920e-01 0.1562995 0.63277546 2.109251e-01